organisation de la structure des codes
