+++ ihm/ihm_6x6
@@ -1,0 +1,104 @@
+import tkinter as tk
+
+class ihm_6x6:
+    
+    def __init__(self):
+        self.taille = 0
+        
+    def grille6x6():
+        global taille, text_id_message, text_ids, grid_values, rows, cols, cell_size, canvas, roots, page1_ouverte, root, offset_x, offset_y
+        taille =0
+        # Définition des paramètres graphique
+        if page1_ouverte == False :
+            roots.destroy()
+        page1_ouverte = True
+        root = tk.Tk()
+        root.title("Grille 6x6")
+
+        # Création du canvas pour dessiner la grille
+        canvas = tk.Canvas(root, width=8*50*1.9, height=8*50*1.3)
+        canvas.pack()
+
+
+        # creation des boutons
+        button_clear = tk.Button(root, text="Réinitialiser", command=clear,width=15, height=3)
+        button_verif = tk.Button(root, text="Vérifier", command=verifier,width=15, height=3)
+        button_valider = tk.Button(root, text="Valider", command=valider,width=15, height=3)
+        button_home = tk.Button(root, text="Accueil", command=home,width=7, height=2)
+        # Placement du bouton dans la fenêtre
+        button_clear.place(x=530, y=100)
+        button_verif.place(x=530, y=200)
+        button_valider.place(x=530, y=300)
+        button_home.place(x=710, y=0)
+        # Dimensions de la grille
+        rows, cols = 6, 6
+        cell_size = 60  # Taille des cellules en pixels
+
+        text_ids = [[None for _ in range(cols)] for _ in range(rows)]
+        grid_values = [["" for _ in range(cols)] for _ in range(rows)]
+        text_id_message = None
+        grid_values1 = [["" for _ in range(cols)] for _ in range(rows)]
+        # Définir les décalages pour centrer la grille
+        offset_x = 50  # Décalage horizontal
+        offset_y = 50  # Décalage vertical
+
+        # Dessin des lignes horizontales et verticales
+        for i in range(rows + 1):
+            canvas.create_line(offset_x, offset_y + i * cell_size, offset_x + cols * cell_size, offset_y + i * cell_size)  # Lignes horizontales
+
+        for j in range(cols + 1):
+            canvas.create_line(offset_x + j * cell_size, offset_y, offset_x + j * cell_size, offset_y + rows * cell_size)  # Lignes verticales
+
+        canvas.bind("<Button-1>", on_click)
+        canvas.create_text(590, 40, text="TAKUZU grille 6x6", font=('Helvetica', 20), fill="black")
+
+        canvas.create_text(80, 440, text="Message", font=('Helvetica', 10), fill="black")
+        canvas.create_rectangle(50, 450, 500, 480, outline="black", width=1, fill="")
+        #numérotation ligne
+        canvas.create_text(40, 60, text="1", font=('Helvetica', 10), fill="black")
+        canvas.create_text(40, 120, text="2", font=('Helvetica', 10), fill="black")
+        canvas.create_text(40, 180, text="3", font=('Helvetica', 10), fill="black")
+        canvas.create_text(40, 240, text="4", font=('Helvetica', 10), fill="black")
+        canvas.create_text(40, 300, text="5", font=('Helvetica', 10), fill="black")
+        canvas.create_text(40, 360, text="6", font=('Helvetica', 10), fill="black")
+
+        #numérotation colonne
+        canvas.create_text(60, 40, text="1", font=('Helvetica', 10), fill="black")
+        canvas.create_text(120, 40, text="2", font=('Helvetica', 10), fill="black")
+        canvas.create_text(180, 40, text="3", font=('Helvetica', 10), fill="black")
+        canvas.create_text(240, 40, text="4", font=('Helvetica', 10), fill="black")
+        canvas.create_text(300, 40, text="5", font=('Helvetica', 10), fill="black")
+        canvas.create_text(360, 40, text="6", font=('Helvetica', 10), fill="black")
+
+        
+        # Remplir le canvas avec les valeurs générées dans grid_values
+        text_ids[0][3] = "Rempli"
+        canvas.create_text(offset_x+3 * cell_size + cell_size // 2,offset_y+ 0 * cell_size + cell_size // 2, text=0, font=('Helvetica', 12), fill="black")
+        grid_values[0][3]=0
+        text_ids[1][0] = "Rempli"
+        canvas.create_text(offset_x+0 * cell_size + cell_size // 2,offset_y+ 1 * cell_size + cell_size // 2, text=1, font=('Helvetica', 12), fill="black")
+        grid_values[1][0]=1
+        text_ids[1][2] = "Rempli"
+        canvas.create_text(offset_x+2 * cell_size + cell_size // 2,offset_y+ 1 * cell_size + cell_size // 2, text=1, font=('Helvetica', 12), fill="black")
+        grid_values[1][2]=1
+        text_ids[2][4] = "Rempli"
+        canvas.create_text(offset_x+4 * cell_size + cell_size // 2,offset_y+ 2 * cell_size + cell_size // 2, text=0, font=('Helvetica', 12), fill="black")
+        grid_values[2][4]=0
+        text_ids[3][2] = "Rempli"
+        canvas.create_text(offset_x+2 * cell_size + cell_size // 2,offset_y+ 3 * cell_size + cell_size // 2, text=1, font=('Helvetica', 12), fill="black")
+        grid_values[3][2]=1
+        text_ids[3][3] = "Rempli"
+        canvas.create_text(offset_x+3 * cell_size + cell_size // 2,offset_y+ 3 * cell_size + cell_size // 2, text=1, font=('Helvetica', 12), fill="black")
+        grid_values[3][3]=1
+        text_ids[3][5] = "Rempli"
+        canvas.create_text(offset_x+5 * cell_size + cell_size // 2,offset_y+ 3 * cell_size + cell_size // 2, text=0, font=('Helvetica', 12), fill="black")
+        grid_values[3][5]=0
+        text_ids[4][1] = "Rempli"
+        canvas.create_text(offset_x+1 * cell_size + cell_size // 2,offset_y+ 4 * cell_size + cell_size // 2, text=0, font=('Helvetica', 12), fill="black")
+        grid_values[4][1]=0
+        text_ids[5][0] = "Rempli"
+        canvas.create_text(offset_x+0 * cell_size + cell_size // 2,offset_y+ 5 * cell_size + cell_size // 2, text=1, font=('Helvetica', 12), fill="black")
+        grid_values[5][0]=1
+
+        for i in range(6) :
+            print(grid_values[i])
